pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["serialscope"]

[project]
name = "serialscope"
version = "0.1.0"
description = "A Modular Serial Telemetry & Debug Framework for Embedded Systems"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "JP", email = "jp@example.com"}
]
keywords = ["serial", "debug", "telemetry", "embedded", "monitoring", "oscilloscope"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Debuggers",
    "Topic :: System :: Hardware",
]

dependencies = [
    "pyserial>=3.5",
    "rich>=13.7.0",
    "textual>=0.52.0",
]

[project.optional-dependencies]
plotting = ["matplotlib>=3.8.0"]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.0",
    "types-pyserial>=3.5.0.20240106",
]

[project.scripts]
serialscope = "serialscope.main:main"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=serialscope",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--strict-markers",
    "-p", "no:launch_testing",
    "-p", "no:launch",
    "-p", "no:launch_testing_ros_pytest_entrypoint",
    "-p", "no:ament_xmllint",
    "-p", "no:ament_pep257",
    "-p", "no:ament_copyright",
    "-p", "no:ament_lint",
    "-p", "no:ament_flake8",
]

[tool.coverage.run]
source = ["serialscope"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
